Initial extraction of SpecSnitch
Setup Guard and RSpec.

Simple integration test.

Simple README instructions

Finally settled on SpecSnitch as name after going through: RSpecSnitch
and Snitch. Snitch would have been nice but there is already a gem with
that name, and RSpec snitch makes it seem like this gem is a module of
RSpec. Besides, we might want to support TestCase in the future.